# Various performance improvements