This PR includes a few performance improvements that are very trivial and don't really need their own PR.

Performance improvements and other notes can be found in the commit messages.

The only change that isn't completely trivial is 99f8cd6326a98f40c49b32e39cd8a0cb3359508e, but that's really just Char#each_byte glued together with String.new.