@wontruefree 2021-12-14
since these are performance improvements do you have any benchmarks?