@BlobCodes 2021-12-14
> since these are performance improvements do you have any benchmarks?

I had benchmarks (the results are in the git commit message), but I don't have them anymore.