@j8r 2021-12-14
Yes, stating x% spees improvement without the actual code is quite meaningless.