@j8r 2021-12-14
Yes, stating x% speed improvement without the actual code is quite meaningless.