@BlobCodes 2021-12-14
[benchmark.zip](https://github.com/crystal-lang/crystal/files/7715288/benchmark.zip)
Benchmarks.
These benchmarks only test against one input (not cherrypicked), but it's probably better to test against different inputs.