@wontruefree 2021-12-15
here is an initial run for people on this thread. Improvements range from 1.03 times improvements to 2.61 times improvement.  I have run these a few times and Blowfish has a 0.01 variance and Char#to_s has a 0.3 variance.

```
$ ./script 
Benchmark 1: Blowfish (02a94043c2d7eb3767dbbfd5e9813d29f521a3ca)
old  18.48  ( 54.12ms) (± 4.12%)  5.13kB/op   1.03× slower
new  19.01  ( 52.61ms) (± 3.91%)  5.02kB/op        fastest
Benchmark 2: URI#request_target (9095f9bc670b693bc3eabb665ddfd31eefbdfea1)
old  11.58M ( 86.36ns) (± 6.03%)  224B/op   1.25× slower
new  14.46M ( 69.16ns) (± 6.93%)  144B/op        fastest
Benchmark 3: URI#resolve_path (e91ac9742385260fbdbc02b61ba24720792be775)
old   1.85M (540.38ns) (± 5.34%)  624B/op   1.22× slower
new   2.26M (442.15ns) (± 4.60%)  448B/op        fastest
Benchmark 4: URI#relativize_path (341d74943ab0b0f363284a395bb385f87e92d0a5)
old   1.22M (817.47ns) (± 6.51%)  916B/op   1.14× slower
new   1.40M (716.56ns) (± 5.10%)  805B/op        fastest
Benchmark 5: Char#to_s (99f8cd6326a98f40c49b32e39cd8a0cb3359508e)
old  37.87M ( 26.41ns) (± 5.86%)  48.0B/op   2.61× slower
new  98.92M ( 10.11ns) (± 6.88%)  16.0B/op        fastest
Benchmark 6: HTTP::Cookie#to_cookie_header (0877acf7c72c7c17a3818ae37a9ebfeb6f7037dc)
old  11.59M ( 86.26ns) (± 6.37%)  208B/op   1.24× slower
new  14.42M ( 69.37ns) (± 7.18%)  128B/op        fastest```