@straight-shoota 2021-12-15
I'm pretty sure we want individual PRs for individual changes. Grouping related changes in different locations is fine. For example the different optimizations passing an initial size to `String.build`.
But please don't put unrelated changes in unrelated locations into a single PR. Having more PRs doesn't cost anything, but it helps to keep the discussions separate and the commit history organized. Thanks.

---
It appears the majority of changes are related to `String.build`, so we could keep them all in this PR and revert all the others, creating separate PRs for them. But I think it's better to create new individual PRs and close this one.