@BlobCodes 2021-12-15
Cherry-picked into #11592, #11593 and #11594